requirements.txt
psycopg2-binary>=2.9.0
requests>=2.31.0
pytz>=2024.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
